worker/requirements.gpu.txt
# Build dependencies are installed in Dockerfile
# Cython and numpy installed first

# Core
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Deep Learning - GPU
# PyTorch CUDA wheels are hosted on the official PyTorch index.
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.2+cu121
torchvision==0.16.2+cu121
--extra-index-url https://pypi.nvidia.com

# YOLO/TensorRT
ultralytics>=8.3.0,<9
tensorrt_bindings==8.6.1
tensorrt_libs==8.6.1

# OCR
easyocr==1.7.0
Pillow==10.1.0

# Computer Vision
opencv-python==4.8.1.78
scipy==1.11.4

# Tracking packages installed separately in Dockerfile
# filterpy, scikit-image, cython-bbox, lap

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
Levenshtein==0.23.0
rapidfuzz==3.5.2

# Monitoring
flower==2.0.1
prometheus-client==0.19.0
